LayoutTests/fast/css/parsing-opacity-expected.txt

Test parsing and computed style of opacity

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS innerStyle("opacity", "0") is "0"
PASS innerStyle("opacity", "0.5") is "0.5"
PASS innerStyle("opacity", "1") is "1"
PASS innerStyle("opacity", "1000") is "1000"
PASS innerStyle("opacity", "-400") is "-400"
PASS innerStyle("opacity", "20%") is "0.2"
PASS innerStyle("opacity", "-600%") is "-6"
PASS innerStyle("opacity", "700%") is "7"
PASS innerStyle("opacity", "2.5e-1") is "0.25"
PASS innerStyle("opacity", "2.5e1%") is "0.25"
PASS innerStyle("opacity", "100.25%") is "1.0025"
PASS computedStyle("opacity", "0") is "0"
PASS computedStyle("opacity", "0.5") is "0.5"
PASS computedStyle("opacity", "1") is "1"
PASS computedStyle("opacity", "1000") is "1"
PASS computedStyle("opacity", "-400") is "0"
PASS computedStyle("opacity", "25%") is "0.25"
PASS computedStyle("opacity", "-600%") is "0"
PASS computedStyle("opacity", "700%") is "1"
PASS computedStyle("opacity", "2.5e-1") is "0.25"
PASS computedStyle("opacity", "2.5e1%") is "0.25"
PASS computedStyle("opacity", "100.25%") is "1"
PASS innerStyle("opacity", "2px") is ""
PASS innerStyle("opacity", "auto") is ""
PASS innerStyle("opacity", "none") is ""
PASS innerStyle("opacity", "'str'") is ""
PASS successfullyParsed is true

TEST COMPLETE
